requirements.txt
notion-client==2.2.1
pyyaml==6.0.1
python-dotenv==1.0.1